requirements.txt
asgiref==3.7.2
astroid==2.9.3
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.6
Django==4.2.9
django-filter==23.5
djangorestframework==3.14.0
djongo==1.3.6
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.12.0
Flask==2.1.2
idna==3.4
importlib-metadata==4.11.4
ipykernel==6.9.1
ipython==8.1.1
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mysql==0.0.3
mysqlclient==2.2.1
nest-asyncio==1.5.4
numpy==1.21.2
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
openpyxl==3.0.10
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==3.5.1
prompt-toolkit==3.0.28
pure-eval==0.2.2
pycryptodome==3.15.0
Pygments==2.11.2
pylint==2.12.2
pymongo==4.1.0
pyparsing==2.4.7
PyPDF2==2.10.6
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
pywin32==303
pyzmq==22.3.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.2.0
toml==0.10.2
tornado==6.1
tqdm==4.66.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2021.5
Unidecode==1.3.6
urllib3==2.0.3
virtualenv==20.23.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.13.3
xlrd==2.0.1
XlsxWriter==3.0.3
xlutils==2.0.0
xlwt==1.3.0
zipp==3.8.0
